Add WorkRoll.js
